Classes for Encodings
Beginning of class declarations for encodings, as well as both Fermionic and Spin Hamiltonians. The end goal will be to have utility functions such that you can put in a system size, an encoding, and either a spin or fermionic Hamiltonian, and get all info about the Majoranas, how to write Paulis, and both the spin and fermionic hamiltonians corresponding to that encoding.